requirements.txt
boto3
pillow
dotenv